Após fazer a classe de teste

+++ src/main/java/tccmaven/GeraArquivoARFF.java
@@ -9,10 +9,9 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStreamWriter;
-import java.lang.reflect.Field;
 import java.text.NumberFormat;
+import java.util.ArrayList;
 import java.util.Date;
-import java.util.Map;
 import java.util.SortedSet;
 import java.util.TreeSet;
 
@@ -21,35 +20,43 @@ import java.util.TreeSet;
  * @author Jean-NoteI5
  */
 public class GeraArquivoARFF {
-    
-    private Map<Date, Parametros> parametros;
-    private String ativo;
-    private final String nomeARFF = ".ARFF";
-    
-    public GeraArquivoARFF(Map<Date, Parametros> linhaArquivo, String ativo) {
-        this.parametros = linhaArquivo;
-        this.ativo = ativo;
+
+    private Parametros parametros;
+    private final String extARFF = ".ARFF";
+
+    public GeraArquivoARFF(Parametros parametros) {
+        this.parametros = parametros;
     }
 
     //Gera arquivo ARFF
     public void geraArquivo() throws GeraArquivoARFFException {
-        
-        
-        try {
-            //Abre o arquivo
 
-            File file = new File(LeituraProperties.getInstance().leituraProperties("prop.diretorioARFF") + ativo + nomeARFF);
+
+        try {
+            
+            String diretorio = LeituraProperties.getInstance().leituraProperties("prop.diretorioARFF");
+            
+            File dir = new File(diretorio);
+            //Se o diretório existe
+            if (!dir.exists()){
+                Log.loga("O diretório " + dir.getAbsolutePath() + " para criação do arquivo ARFF não existe");
+                throw new GeraArquivoARFFException("Não foi possível gerar o arquivo ARFF");
+            }
+            
+            //Abre o arquivo
+            File file = new File(diretorio + parametros.getAtivo() + extARFF);
             Log.loga("Arquivo ARFF: " + file.getAbsolutePath());
+            
             FileOutputStream arquivoGravacao = new FileOutputStream(file);
             OutputStreamWriter strWriter = new OutputStreamWriter(arquivoGravacao);
             BufferedWriter writer = new BufferedWriter(strWriter);
-            
-            
+
+
             writer.write("% This is a dataset obtained from the YAHOO FINANCES. Here is the included description:");
             writer.newLine();
             writer.write("%");
             writer.newLine();
-            writer.write(new String("% The data provided are daily stock prices from #INICIO# through #FIM#, for #ATIVO#.").replaceAll("#INICIO#", LeituraProperties.getInstance().leituraProperties("prop.DataIni")).replaceAll("#FIM#", LeituraProperties.getInstance().leituraProperties("prop.DataFim")).replaceAll("#ATIVO#", ativo));
+            writer.write(new String("% The data provided are daily stock prices from #INICIO# through #FIM#, for #ATIVO#.").replaceAll("#INICIO#", LeituraProperties.getInstance().leituraProperties("prop.DataIni")).replaceAll("#FIM#", LeituraProperties.getInstance().leituraProperties("prop.DataFim")).replaceAll("#ATIVO#", parametros.getAtivo()));
             writer.newLine();
             writer.write("%");
             writer.newLine();
@@ -63,19 +70,15 @@ public class GeraArquivoARFF {
             writer.newLine();
             writer.newLine();
 
-            //Obtém os campos da classe Parâmetros
-            Field field[] = Parametros.class.getDeclaredFields();
 
-            //Varre os campos existentes na classe para exportar o cabeçalho
-            for (int i = 0; i < field.length; i++) {
-                //Ignora o campo de data
-                if (field[i].getName().equals("date")) {
-                    continue;
-                }
-                writer.write("@attribute " + field[i].getName() + " numeric");
+            //Obtém a lista de parâmetros existentes nos parâmetros
+            SortedSet<String> nomesParametros = new TreeSet<>(parametros.getNomeParametros());
+
+            for (String nomeParametro : nomesParametros) {
+                writer.write("@attribute " + nomeParametro + " numeric");
                 writer.newLine();
             }
-            
+
             writer.newLine();
             writer.write("@data");
             writer.newLine();
@@ -83,46 +86,47 @@ public class GeraArquivoARFF {
 
 
             //Ordenado as chaves do HashMap
-            SortedSet<Date> chaves = new TreeSet<>(parametros.keySet());
+            SortedSet<Date> chaves = new TreeSet<>(parametros.getParametros().keySet());
 
             //Percorre as chaves do array
             for (Date chave : chaves) {
                 StringBuilder linha = new StringBuilder();
-                for (int j = 0; j < field.length; j++) {
+                //Parametros existentes na data lida na chave
+                ArrayList<Parametro> aux = parametros.getParametros().get(chave);
 
-                    //Ignora o campo de data
-                    if (field[j].getName().equals("date")) {
-                        continue;
+                //Exporta na mesma ordenação dos parâmetros
+                for (String nomeParametro : nomesParametros) {
+                    //Varre os parâmetros existentes
+                    for (int i = 0; i < aux.size(); i++) {
+                        //Se for parâmetro correto
+                        if (aux.get(i).getDescrição().equals(nomeParametro)) {
+                            /* 
+                             * Realiza a formatação do número para evitar que ele seja representado 
+                             * como notação cientifica quando o seu valor for muito grande
+                             */
+                            Double edita = aux.get(i).getValor();
+                            NumberFormat f = NumberFormat.getInstance();
+                            f.setGroupingUsed(false);
+                            f.setMaximumFractionDigits(2);
+                            String editado = f.format(edita.doubleValue()).replaceAll(",", ".");
+                            linha.append(editado);
+                            linha.append(",");
+                        }
                     }
-
-                    /* 
-                     * Realiza a formatação do número para evitar que ele seja representado 
-                     * como notação cientifica quando o seu valor for muito grande
-                     */
-                    Double edita = field[j].getDouble(parametros.get(chave));
-                    
-                    
-                    
-                    NumberFormat f = NumberFormat.getInstance();
-                    f.setGroupingUsed(false);
-                    f.setMaximumFractionDigits(2);
-                    String editado = f.format(edita.doubleValue()).replaceAll(",", ".");
-                    linha.append(editado);
-                    if (j < field.length - 1) {
-                        linha.append(",");
-                    }
-                    
                 }
+                //Remove a última vírgula da literal
+                linha.delete(linha.length() - 1, linha.length());
+                //Exporta a linha para o arquivo
                 writer.write(linha.toString());
                 writer.newLine();
-                
+
             }
             //Fecha o arquivo
             writer.close();
-            
-        } catch (IOException | IllegalArgumentException | IllegalAccessException ex) {
+
+        } catch (IOException | IllegalArgumentException ex) {
             throw new GeraArquivoARFFException("Ocorreu erro no momento de gerar o arquivo ARFF", ex);
         }
-        
+
     }
 }
